Livonia Pit Set-up 	- adjusted distance for auton three tote back 	- adjusted roller speed for auton three tote back 	- when setting 6 tote stack, change wrist position

+++ src/Arm.cpp
@@ -228,7 +228,7 @@ void Arm::shoulderSetSetpoint(int point){
 	m_shoulderPid->SetSetpoint(point);
 }
 
-void Arm::wristSetSetpoint(int point){
+void Arm::wristSetSetpoint(float point){
 	m_wristPid->SetSetpoint(point);
 }
